+++ software/src/modules/front_panel/front_panel.h
@@ -97,7 +97,7 @@ private:
 
     void check_flash_metadata();
 
-    String watt_value_to_display_string(const float watt);
+    String watt_value_to_display_string(const int32_t watt);
     String watt_hour_value_to_display_string(const uint32_t kilo_watt_hour);
     String price_value_to_display_string(const int32_t price);
